requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
numpy>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pypdf>=3.0.0

